combine Wasm and PE
Now, we'll first run Partial Eval on the queries that the engine cares
about, and then compile the result into wasm modules.

As it turns out, the input format for wasm matters, and it's BETTER to
pass RAW input -- the wasm sdk in OPA will call json.Marshal on it,
and you'll profit from that being faster than String() called on an
ast.Value.

Note: we're using OPA@main here, so take care of the underscore import.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>